+++ anrtd-client/src/pages/EditToDo/EditToDo.tsx
@@ -1,17 +1,12 @@
 import React, { useCallback } from 'react';
 import * as api from '../../api';
 import { appPaths, useToDoId } from '../../AppRoutes';
-import ToDoForm, { ToDoFormComponent, ToDoFormValues } from '../../components-todos/Form';
+import { ToDoFormValues } from '../../components-todos/Form';
 import ApiResponseWrapper from '../../components/ApiResponseWrapper';
-import AppPageHeading from '../../components/AppPageHeading';
-import AppFormikSubmitButton from '../../components/AppForm/AppFormikSubmitButton';
-import AppButton, { AppButtons } from '../../components/AppButton';
-import TShirtIcon from '../../components/TShirtIcon';
-import useStyles from './EditToDoStyles';
+import EditToDoPageComp from './EditToDoComp';
 
 const EditToDoPage: React.FC = () => {
     const toDoId = useToDoId();
-    const classes = useStyles();
     const fetchToDo = useCallback(() => api.getToDoById(toDoId), [toDoId]);
     const {
         isLoading,
@@ -37,28 +32,13 @@ const EditToDoPage: React.FC = () => {
                 isFetching={isLoading}
                 isError={isError}
             >
-                <ToDoForm
-                    onSubmit={handleUpdateToDo}
-                    initialValues={toDo}
-                >
-                    {formikProps => (
-                        <>
-                            <div className={classes.heading}>
-                                {(formikProps.values.tShirtSize || formikProps.values.tShirtSize === 0) && (
-                                    <TShirtIcon size={formikProps.values.tShirtSize} />
-                                )}
-                                <AppPageHeading gutterBottom={false}>Edit ToDo #{toDoId}</AppPageHeading>
-                            </div>
-                            <ToDoFormComponent
-                                {...formikProps}
-                            />
-                            <AppButtons>
-                                <AppButton linkPath={appPaths.toDos}>Back</AppButton>
-                                <AppFormikSubmitButton>Save</AppFormikSubmitButton>
-                            </AppButtons>
-                        </>
-                    )}
-                </ToDoForm>
+                {toDo && (
+                    <EditToDoPageComp
+                        toDo={toDo}
+                        handleUpdateToDo={handleUpdateToDo}
+                        backLinkPath={appPaths.toDos}
+                    />
+                )}
             </ApiResponseWrapper>
         </>
     );
